+++ modules/Content/views/collection/item.php
@@ -138,7 +138,7 @@
             <kiss-container>
                 <div class="kiss-flex kiss-flex-middle">
                     <div class="kiss-button-group" v-if="item._id">
-                        <a class="kiss-button" href="<?=$this->route("/content/collection/item/{$model['name']}")?>">
+                        <a class="kiss-button" :href="$route(`/content/collection/item/${model.name}`)">
                             <?=t('Create new item')?>
                         </a>
                         <a class="kiss-button" :href="$route(`/content/collection/clone/${model.name}/${item._id}`)">