+++ modules/Content/views/tree/item.php
@@ -137,8 +137,11 @@
             <kiss-container>
                 <div class="kiss-flex kiss-flex-middle">
                     <div class="kiss-button-group" v-if="item._id">
-                        <a class="kiss-button" href="<?=$this->route("/content/tree/item/{$model['name']}")?>">
+                        <a class="kiss-button" :href="$route(`/content/tree/item/${model.name}`)">
                             <?=t('Create new item')?>
+                        </a>
+                        <a class="kiss-button" :href="$route(`/content/tree/clone/${model.name}/${item._id}`)">
+                            <?=t('Clone item')?>
                         </a>
                     </div>
                     <div class="kiss-flex-1"></div>